requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pdfplumber==0.11.4
pandas==2.2.2
pyarrow==17.0.0
sentence-transformers==3.1.1
chromadb==0.5.7
fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit==1.39.0
ollama==0.3.3
pydantic-settings==2.5.2
tenacity==9.0.0
pytest==8.3.3
mypy==1.11.2
black==24.8.0
python-dotenv==1.0.1
